requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
cffi==1.12.2
constantly==15.1.0
cryptography==2.6.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
pycparser==2.19
PyHamcrest==1.9.0
pyjson==1.3.0
pyOpenSSL==19.0.0
python-dotenv==0.10.1
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
zope.interface==4.6.0
